requirements_dev.txt
pytest
mockredispy
bumpversion
